Parameters to couple of methods to allow caller decide the order and limit
Added parameters (with default values to those that used to be) that allow caller to pass order by and limit parameters when getting list of test sessions. This is needed for RSS since it needs to publish by created_at instead of tested_at, and it needs not to show all found sessions. The limit is set to 10 now.