BMC#13552 proposal
Restrict the allowed characters in test type and hardware so that user cannot break navigation/functionality of QA reports. This proposal allows word characters (a-zA-Z0-9), spaces and special characters , ; : ( ) Dot is not allowed as it did break the functionality. I'd suppose editing routes would allow using of dot as well.