Do not set any color width to zero percent if value is not zero
When a certain filtered list would contain a report that has a lot of cases and a report that has only few cases the html graph would not necessarily display correct colors. An example scenario is a set with 100 cases and another set with 5 passed cases and 1 failed case. The graph in this scenario will be all green even though there's a failed case.