BMC#9808 Ability to remove test cases
With this merge request test cases can be removed in edit mode. This will also update has_nft and has_non_nft things for sets and session. If a test set has no cases after it is removed the set will also be destroyed.

This does not update case counts etc. dynamically in the edit mode but I suppose that's not very important? Also if a set is deleted it will show on the page while in edit mode but after leaving edit mode it naturally is not visible anymore.