Implement the feature FEA#13526
FEA#13526: Request to get the detailed test results as CSV file from an existed report

I add a link(Download as CSV) in report view. After press it,
the existed report test result is able to be exported as CSV file, and we can upload
it as a new report.

Signed-off-by: Huang Rui <rui.r.huang@intel.com>