Implement BMC#13976 Build-ID should be managed
A build ID could be committed (optionally) when creating a new report.
The buildID should be creatable, editable, and selectable.

Add a Build_id column into meego_test_session, for comparing the test report of
the same build over different hardware (BMC#13975 and BMC#13979).
Meanwhile, a maintaince page is also needed for edit/add the buildID

Signed-off-by: Huang Rui <rui.r.huang@intel.com>