Simplified tests and use the latest API values as default
Some of the tests were unnecessary complex for what they were testing. Reuse simpler steps more.

Changed the default_api_opts to use the current API, and added tests for the two previous ones. I don't think using the term "new api" works in a long run when we're already running the 3rd version of it.